refactor(supabase): extract validation and error handling into focused files
Split supabase.go into three files to improve maintainability:
- errors.go: PostgREST error response parsing and error type mapping
- validate.go: table/filter/column validation, parseAndValidate generic,
  and new returningSelect() DRY helper

Updated delete.go, insert.go, and update.go to use returningSelect()
instead of duplicating the returning/select pattern inline.
