+++ connectors/supabase/update.go
@@ -53,12 +53,7 @@ func (a *updateAction) Execute(ctx context.Context, req connectors.ActionRequest
 
 	q := url.Values{}
 
-	// Column selection for returned rows.
-	ret := params.Returning
-	if ret == "" {
-		ret = "*"
-	}
-	q.Set("select", ret)
+	returningSelect(q, params.Returning)
 
 	// Apply filters to scope the update.
 	if err := applyFilters(q, params.Filters); err != nil {
